adding first iteration of code
The work so far just crawls the forum for a hard coded number of pages,
and saves the text in posts, for insertion into a nltk text class. This
can then be processed to extract data. NOTE that there is currently a
bug, because I tried to append a list to a list, and it actually inserts
the list.